Move auth routes to user routes
